requirements.txt
gradio==4.44.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
tenacity==9.0.0